ci: actually configure publish from CI
There isn't any way as far as I can tell to "only" configure an NPM token when publishing
(and not when installing):
- `publishConfig` isn't actually implemented as documented in yarn, so you can't specify
  `_authToken` that way
- you get an error if you configure `_authToken` in `.npmrc` but the env var you're
  reading from isn't defined sometimes (i.e. when installing)
- `always-auth = false` would do what we want (not authenticate when installing), but this
  is moot because of the above points

The only viable approach is to generate a different `.npmrc` on the fly when publishing
(or have two committed `.npmrc`s). That seems gross relative to the actual value of any of
this. So forget it.